[fix]: Flexibilidade na escolha da chave primária

+++ crud.py
@@ -1,23 +1,28 @@
 import pymongo
+import structures
 
 class CrudDB:    
-    def __init__(self, local_condition, db_name, collection_name, primary_key="", url=""):
+    def __init__(self, local_condition, db_name, collection_name, primary_key="cpf", url=""):
         if local_condition is True:
             self.client = pymongo.MongoClient("localhost", 27017)
         else:
             self.client = pymongo.MongoClient(url)
 
+        default_key = "cpf"
+        self.primary_key = primary_key
+
         self.database = self.client[db_name]
         self.collection = self.database[collection_name]
 
-        if self.collection is not None and primary_key != "":
-            self.collection.create_index(primary_key)    
-
-        #self.content = []
+        if self.collection is not None:
+            if self._verify_key(primary_key): # A inserção de nenhum valor vai cair aqui por default
+                self.collection.create_index(primary_key)
+            else:
+                self.collection.create_index(default_key)   
 
     # Passar esse método para a outra classe depois
     def _format_content(self, content, divisor=True):
-        format_keys = {"cpf": "CPF", "nome": "Primeiro Nome", "idade": "Idade (em anos)", "email": "E-mail"}
+        format_keys = structures.DOCUMENT
         for key in content.keys():
             old_key = key
             if key in format_keys.keys():
@@ -26,6 +31,13 @@ class CrudDB:
         if divisor is True:    
             print("------------------------")      
 
+    # Verificar presença de chave no "structures.py"
+    def _verify_key(self, key):
+        if key in structures.DOCUMENT:
+            return True
+        return False
+
+    # Alterar para utilizar o novo método find_by_primary_key
     def read_data(self, query={}):
         try:
             content = self.collection.find(query)
@@ -44,31 +56,24 @@ class CrudDB:
                       
 
     # Atualizar para adaptar ao parameter primary_key
-    def find_by_cpf(self, cpf, no_print=False):
+    def find_by_primary_key(self, key):
         # Return values:
         # Status + conteúdo
-        query = {"cpf": cpf}
+        query = {self.primary_key: key}
         content = self.collection.find_one(query)
-        if no_print == False:
-            #print("Resultado da seleção do CPF {0}".format(cpf))
-            #self._format_content(content, divisor=False)
-            return True, content
-        else:
-            if content == None:
-                return False, content
-            return True, content 
+        if content == None:
+            return False, content
+        return True, content 
 
-    def create_customer(self, cpf, name, age, email):
+    def create_customer(self, primary_key, values=structures.DOCUMENT):
         # False --> Não foi possível executar a criação
-        has_already_cpf = self.find_by_cpf(cpf, no_print=True)
-        if has_already_cpf == False:
-            document = {
-        # True  --> Criação feita com sucesso
-                "cpf": cpf,
-                "nome": name,
-                "idade": age,
-                "email": email
-            }
+        # True  --> Criação feita com sucesso        
+        has_already_register, content = self.find_by_primary_key(primary_key)
+        if has_already_register == False:
+            document = structures.DOCUMENT
+            for key in document.keys():
+                if key in values.keys():
+                    document[key] = values[key]
 
             try:
                 self.collection.insert_one(document)
@@ -78,43 +83,38 @@ class CrudDB:
         return False        
 
 
-    def update_customer_cpf(self, cpf, name="", age="", email=""):
+    def update_customer(self, primary_key, values):
         # True --> Atualização realizada
         # False --> Nenhum valor informado ou falha ao atualizar o registro
-
-        query = {"cpf": cpf}
-        values = {}
-
-        # Adicionando valores no set_values
-        if name != "":
-            values["nome"] = name
-        
-        if age != "":
-            values["idade"] = age
-        
-        if email != "":
-            values["email"] = email
-
         # Nenhuma informação foi dada para ser preenchida
         if values == {}:
             return False
+
+        register_exists, content = self.find_by_primary_key(primary_key)
+        if register_exists:
+            query = {self.primary_key: primary_key}
+
+            set_values = {"$set": values}
+
+            try:
+                self.collection.update_one(query, set_values)
+                return True
+            except:
+                return False
         
-        set_values = {"$set": values}
+        return False
 
-        try:
-            self.collection.update_one(query, set_values)
-            return True
-        except:
-            return False
+    def delete_customer(self, primary_key):
+        register_exists, content = self.find_by_primary_key(primary_key)
+        if register_exists:
+            query = {self.primary_key: primary_key}
 
-    def delete_customer_cpf(self, cpf):
-        query = {"cpf": cpf}
-
-        try:
-            self.collection.delete_one(query)
-            return True
-        except:
-            return False
+            try:
+                self.collection.delete_one(query)
+                return True
+            except:
+                return False
+        return False
 
 #system = CrudDB(local_condition=True, db_name='Clientes', collection_name='Clientes', primary_key="cpf")
